+++ R/DecisionTree.R
@@ -1,9 +1,9 @@
 #' @title A decision tree
 #' @description An R6 class to represent a decision tree model.
-#' @details A class to represent a decision tree. An object contains a tree of
-#' decision nodes, chance nodes and leaf nodes, connected by edges
-#' (either actions or reactions). It inherits from class \code{Arborescence} and
-#' satisfies the following conditions:
+#' @section Constructing the tree:
+#' A \code{DecisionTree} object contains decision nodes, chance nodes and leaf
+#' nodes, connected by edges (either actions or reactions). It inherits from
+#' class \code{Arborescence} and satisfies the following conditions:
 #' \enumerate{
 #' \item{Nodes and edges must form a tree with a single root and
 #' there must be a unique path from the root to each node.
@@ -27,6 +27,17 @@
 #' \item{Each \code{Action} must have a label, and the labels of
 #' \code{Action}s that share a common source endpoint must be unique.}
 #' }
+#' @section Time:
+#' The timing of events is not explicitly modelled in decision trees
+#' (O'Mahony, 2015). \code{rdecision} makes the assumption that all costs are
+#' incurred at time \eqn{t = 0} and that QALYs are gained during the
+#' time intervals defined for each leaf node. Discounting of future costs to
+#' present values is therefore not applicable. Future utilities may be
+#' discounted to present values by setting a discount rate for
+#' each leaf node (in usual circumstances the interval and discount rate should
+#' be the same for each leaf node). The QALYs gained are calculated by
+#' integrating the continuously discounted utility over the interval for each
+#'  leaf node.
 #' @references{
 #'   Briggs A, Claxton K, Sculpher M. Decision modelling for health economic
 #'   evaluation. Oxford, UK: Oxford University Press; 2006.
@@ -40,6 +51,11 @@
 #'   Kaminski B, Jakubczyk M, Szufel P. A framework for sensitivity analysis of
 #'   decision trees. \emph{Central European Journal of Operational Research}
 #'   2018;\bold{26}:135–59, \doi{10.1007/s10100-017-0479-6}.
+#'
+#'   O'Mahony JF, Newall AT, van Rosmalen J. Dealing with time in health
+#'   economic evaluation: methodological issues and recommendations for
+#'   practice. \emph{PharmacoEconomics} 2015;\bold{33}:1255-1268,
+#'   \doi{10.1007/s40273-015-0309-4}.
 #' }
 #' @docType class
 #' @author Andrew J. Sims \email{andrew.sims@@newcastle.ac.uk}
@@ -849,7 +865,8 @@ DecisionTree <- R6::R6Class(
     #' @details For each walk, probability, cost, benefit and utility are
     #' calculated. There is minimal checking of the argument because this
     #' function is intended to be called repeatedly during tree evaluation,
-    #' including PSA.
+    #' including PSA. Discounts are applied to calculate the present value of
+    #' future costs and health effects.
     #' @param W A list of root-to-leaf walks. A walk is a sequence of edges
     #' (actions and reactions), stored as a list. Each walk must start with an
     #' edge whose source is the root node and end with an edge whose target is
@@ -957,10 +974,10 @@ DecisionTree <- R6::R6Class(
     #'       leaving the node}
     #'     \item{\code{Leaf}}{The label of terminating leaf node}
     #'     \item{\code{Probability}}{Probability of traversing the path}
-    #'     \item{\code{Cost}}{Cost of traversing the path}
-    #'     \item{\code{Benefit}}{Benefit of traversing the path}
-    #'     \item{\code{Utility}}{Utility of traversing the path}
-    #'     \item{\code{QALY}}{QALY of traversing the path}
+    #'     \item{\code{Cost}}{Cost of traversing the path x probability}
+    #'     \item{\code{Benefit}}{Benefit of traversing the path x probability}
+    #'     \item{\code{Utility}}{Utility of traversing the path x probability}
+    #'     \item{\code{QALY}}{QALYs gained by those traversing the path}
     #'   }
     #' }
     #'